all planned V1.0 features working

+++ src/controllers/DeviceController.h
@@ -21,7 +21,7 @@ class DeviceController {
   bool enableAlarm_A6;
 
     bool deviceState;
-    int currentCycleCount;
+    byte currentCycleCount;
     int totalCycleCount;
